testtaskDisoftproject/requirements.txt
asgiref==3.5.0
black==21.12b0
click==8.0.3
colorama==0.4.4
Django==4.0.1
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
